style: rename webgl state file for consistency
