+++ src/core/renderable_object.ts
@@ -5,7 +5,7 @@ import { Texture } from "../objects/textures/texture";
 import { TrsTransform } from "../core/transforms";
 import { Shader } from "../renderers/shaders";
 import { Color } from "../core/color";
-import { CullingMode } from "../renderers/WebGLState";
+import { CullingMode } from "../renderers/webgl_state";
 
 export abstract class RenderableObject extends Node {
   public wireframeEnabled = false;
